io: reduce noise using module-level clippy deny
