+++ embedded-io-adapters/src/std.rs
@@ -1,5 +1,10 @@
 //! Adapters to/from `std::io` traits.
 
+#![deny(
+    clippy::missing_trait_methods,
+    reason = "Methods should be forwarded to the underlying type"
+)]
+
 use embedded_io::Error as _;
 
 /// Adapter from `std::io` traits.
@@ -36,10 +41,6 @@ impl<T: ?Sized> embedded_io::ErrorType for FromStd<T> {
     type Error = std::io::Error;
 }
 
-#[deny(
-    clippy::missing_trait_methods,
-    reason = "Methods should be forwarded to the underlying type"
-)]
 impl<T: std::io::Read + ?Sized> embedded_io::Read for FromStd<T> {
     fn read(&mut self, buf: &mut [u8]) -> Result<usize, Self::Error> {
         self.inner.read(buf)
@@ -59,10 +60,6 @@ impl<T: std::io::Read + ?Sized> embedded_io::Read for FromStd<T> {
     }
 }
 
-#[deny(
-    clippy::missing_trait_methods,
-    reason = "Methods should be forwarded to the underlying type"
-)]
 impl<T: std::io::BufRead + ?Sized> embedded_io::BufRead for FromStd<T> {
     fn fill_buf(&mut self) -> Result<&[u8], Self::Error> {
         self.inner.fill_buf()
@@ -73,10 +70,6 @@ impl<T: std::io::BufRead + ?Sized> embedded_io::BufRead for FromStd<T> {
     }
 }
 
-#[deny(
-    clippy::missing_trait_methods,
-    reason = "Methods should be forwarded to the underlying type"
-)]
 impl<T: std::io::Write + ?Sized> embedded_io::Write for FromStd<T> {
     fn write(&mut self, buf: &[u8]) -> Result<usize, Self::Error> {
         match self.inner.write(buf) {
@@ -102,10 +95,6 @@ impl<T: std::io::Write + ?Sized> embedded_io::Write for FromStd<T> {
     }
 }
 
-#[deny(
-    clippy::missing_trait_methods,
-    reason = "Methods should be forwarded to the underlying type"
-)]
 impl<T: std::io::Seek + ?Sized> embedded_io::Seek for FromStd<T> {
     fn seek(&mut self, pos: embedded_io::SeekFrom) -> Result<u64, Self::Error> {
         self.inner.seek(pos.into())
