+++ embedded-io-async/src/impls/blanket.rs
@@ -3,27 +3,23 @@
     reason = "Methods should be forwarded to the underlying type"
 )]
 
-use crate::{BufRead, Read, Seek, SeekFrom, Write};
-use alloc::boxed::Box;
+use embedded_io::{ReadExactError, SeekFrom};
 
-#[cfg_attr(docsrs, doc(cfg(any(feature = "std", feature = "alloc"))))]
-impl<T: ?Sized + Read> Read for Box<T> {
+use crate::{BufRead, Read, Seek, Write};
+
+impl<T: ?Sized + Read> Read for &mut T {
     #[inline]
     async fn read(&mut self, buf: &mut [u8]) -> Result<usize, Self::Error> {
         T::read(self, buf).await
     }
 
     #[inline]
-    async fn read_exact(
-        &mut self,
-        buf: &mut [u8],
-    ) -> Result<(), crate::ReadExactError<Self::Error>> {
+    async fn read_exact(&mut self, buf: &mut [u8]) -> Result<(), ReadExactError<Self::Error>> {
         T::read_exact(self, buf).await
     }
 }
 
-#[cfg_attr(docsrs, doc(cfg(any(feature = "std", feature = "alloc"))))]
-impl<T: ?Sized + BufRead> BufRead for Box<T> {
+impl<T: ?Sized + BufRead> BufRead for &mut T {
     #[inline]
     async fn fill_buf(&mut self) -> Result<&[u8], Self::Error> {
         T::fill_buf(self).await
@@ -35,26 +31,24 @@ impl<T: ?Sized + BufRead> BufRead for Box<T> {
     }
 }
 
-#[cfg_attr(docsrs, doc(cfg(any(feature = "std", feature = "alloc"))))]
-impl<T: ?Sized + Write> Write for Box<T> {
+impl<T: ?Sized + Write> Write for &mut T {
     #[inline]
     async fn write(&mut self, buf: &[u8]) -> Result<usize, Self::Error> {
         T::write(self, buf).await
     }
 
     #[inline]
-    async fn write_all(&mut self, buf: &[u8]) -> Result<(), Self::Error> {
-        T::write_all(self, buf).await
-    }
-
-    #[inline]
     async fn flush(&mut self) -> Result<(), Self::Error> {
         T::flush(self).await
     }
+
+    #[inline]
+    async fn write_all(&mut self, buf: &[u8]) -> Result<(), Self::Error> {
+        T::write_all(self, buf).await
+    }
 }
 
-#[cfg_attr(docsrs, doc(cfg(any(feature = "std", feature = "alloc"))))]
-impl<T: ?Sized + Seek> Seek for Box<T> {
+impl<T: ?Sized + Seek> Seek for &mut T {
     #[inline]
     async fn seek(&mut self, pos: SeekFrom) -> Result<u64, Self::Error> {
         T::seek(self, pos).await
