+++ embedded-io/src/impls/mod.rs
@@ -1,3 +1,4 @@
+mod blanket;
 mod slice_mut;
 mod slice_ref;
